% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evasys.skala.plot.R
\name{barplot_evasys}
\alias{barplot_evasys}
\title{Barplot-Boxplot Hypbrid für die Darstellung von ordinalskalierten Variablen}
\usage{
barplot_evasys(x, tmin, tmax, number = 6)
}
\arguments{
\item{x}{Daten}

\item{tmin}{Beschriftung links}

\item{tmax}{Beschriftung rechts}

\item{number}{Skala (6 für Sechserskala etc.)}
}
\value{
Barplot-Boxplot-Hybrid
}
\description{
Die optimale Chunk-Einstellung hierfür ist: fig.width = 6, fig.height = 1.4
}
\examples{
barplot_evasys(BspDaten$dataSHOWUP$info_ausr_studgang,
               tmin = "stimme gar nicht zu", tmax = "stimme voll zu")


}
