% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bar.R
\name{barplot_freq}
\alias{barplot_freq}
\title{Balkendiagramm der Häufigkeiten}
\usage{
barplot_freq(x, xlab = "")
}
\arguments{
\item{x}{Faktor (die Stufen bestimmen Reihenfolge und Beschriftung der
Balken).}

\item{xlab}{Beschriftung der x-Achse.}
}
\value{
Kein verwertbarer Wert (unsichtbar); die Abbildung wird gezeichnet.
}
\description{
Zeichnet ein senkrechtes Balkendiagramm mit der Häufigkeit jeder Kategorie,
z. B. für Fachsemester oder in Klassen eingeteilte Noten. Wird von
\code{\link[=merge_num]{merge_num()}} und \code{\link[=merge_fachsem]{merge_fachsem()}} verwendet.
}
\examples{
barplot_freq(BspDaten$Plots$fsem, xlab = "Fachsemester")

}
\seealso{
Weitere Grafikfunktionen:
\code{\link[=barplot_scmc]{barplot_scmc()}},
\code{\link[=barplot_sk]{barplot_sk()}},
\code{\link[=boxplot_aggr_sk]{boxplot_aggr_sk()}},
\code{\link[=boxplot_grade]{boxplot_grade()}},
\code{\link[=boxplot_rueck]{boxplot_rueck()}},
\code{\link[=boxplot_wl]{boxplot_wl()}}
}
\concept{grafiken}
