% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.gesnote.R
\name{boxplot.gesnote}
\alias{boxplot.gesnote}
\title{Abbildung der Gesamtnote}
\usage{
\method{boxplot}{gesnote}(x)
}
\arguments{
\item{x}{Daten}
}
\value{
Boxplot der Gesamtnote
}
\description{
Abbildung der Gesamtnote
}
