% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.ruecklauf.R
\name{boxplot.ruecklauf}
\alias{boxplot.ruecklauf}
\title{Boxplot für Rücklaufsabbildung (LVE)}
\usage{
\method{boxplot}{ruecklauf}(x, kennung)
}
\arguments{
\item{x}{Objekt mit Teilnehmendenzahlen}

\item{kennung}{Kennungen}
}
\value{
Boxplot der Rückläufe
}
\description{
Boxplot für Rücklaufsabbildung (LVE)
}
\examples{
boxplot.ruecklauf(BspDaten$dataLVE$Teilnehmer, BspDaten$dataLVE$Kennung)

}
