% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.vert.R
\name{boxplot.vert}
\alias{boxplot.vert}
\title{Vertikaler Boxplot}
\usage{
\method{boxplot}{vert}(x, kennung, dont.know = "")
}
\arguments{
\item{x}{Daten}

\item{kennung}{Kennung}

\item{dont.know}{gibt es eine "weiß nicht"-Option?}
}
\value{
Boxplot
}
\description{
Vertikaler Boxplot
}
