% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.workload.R
\name{boxplot.workload}
\alias{boxplot.workload}
\title{Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden}
\usage{
\method{boxplot}{workload}(
  x,
  skala = c("0h", "1h", "2h", "3h", "4h", "5h", "6h", "7h", "8h", "9h", "10h", "11h",
    "12h", "mehr als\\n12h")
)
}
\arguments{
\item{x}{Daten}

\item{skala}{Skala x-Achse}
}
\value{
Boxplot
}
\description{
Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden
}
\examples{
boxplot.workload(x = BspDaten$Plots$WL)

}
