% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_box.R
\name{boxplot_grade}
\alias{boxplot_grade}
\title{Abbildung der Gesamtnote}
\usage{
boxplot_grade(x)
}
\arguments{
\item{x}{(aggregierte) Daten für den Boxplot}
}
\value{
Boxplot der Gesamtnote
}
\description{
Abbildung der Gesamtnote
}
\examples{
boxplot_grade(BspDaten$Plots$grade)

}
