% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_box.R
\name{boxplot_wl}
\alias{boxplot_wl}
\title{Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden}
\usage{
boxplot_wl(
  x,
  skala = c("0h", "1h", "2h", "3h", "4h", "5h", "6h", "7h", "8h", "9h", "10h", "11h",
    "12h", "mehr\\nals 12h")
)
}
\arguments{
\item{x}{Daten}

\item{skala}{Skala x-Achse}
}
\value{
Boxplot
}
\description{
Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden
}
\examples{
boxplot_wl(x = BspDaten$Plots$WL)

}
