% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.table.stat.R
\name{bsp.table.stat}
\alias{bsp.table.stat}
\title{Beispiel einer Statistik-Tabelle}
\usage{
bsp.table.stat(all = TRUE)
}
\arguments{
\item{all}{Soll die Tabelle Fragen enthalten? Dann TRUE, sonst FALSE}
}
\value{
Eine Tabelle mit den Spalten "Häufigkeit", "Mittelwert", "Standardabweichung",
"Median", "kleinster beob. Wert", "größter beob. Wert"
}
\description{
Beispiel einer Statistik-Tabelle
}
\examples{
bsp.table.stat()

}
