% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BspDaten}
\alias{BspDaten}
\title{Beispiel-Daten}
\format{
\subsection{\code{BspDaten}}{

Eine Liste mit fünf Elementen

\describe{
\item{dataLVE}{Daten der LVE 2024}
\item{dataSHOWUP}{Daten der SHOWUP 2024/25}
\item{Plots}{Daten für die Testung der Bar- und Boxplotfunktionen}
\item{Tabellen}{Daten für die Testung der Tabellenfunktionen}
\item{pInfo}{Exemplarische personalised-Info-Tabelle}
}
}
}
\usage{
BspDaten
}
\description{
Dieses Objekt enthält mehrere Beispieldaten, die in den Beispielen der
Dokumentation genutzt werden. Die Daten entstamme der LVE 2024 und der SHOWUP 2024/25.
}
\examples{
BspDaten$dataLVE$FachSemN |> hist()
}
\keyword{datasets}
