% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{change.analysis.defaults}
\alias{change.analysis.defaults}
\title{Funktion um Einstellungsvariablen anzupassen}
\usage{
change.analysis.defaults(...)
}
\arguments{
\item{...}{Argumente, die die Einstellungen ändern sollen, z. B.
\code{color.bars = "red"} oder \code{show.plot.sc = FALSE}.
Die Namen der Argumente müssen mit den Namen der Einstellungsvariablen übereinstimmen.}
}
\description{
Funktion um Einstellungsvariablen anzupassen
}
\examples{

# Hier wird die Farbe der Balken auf rot geändert und eingestellt,
# dass keine SC-Plots gezeigt werden sollen

change.analysis.defaults(color.bars = "red", show.plot.sc = FALSE)

#Eine Überprüfung zeigt, dass die Änderungen erfolgreich waren

setanalysis_defaults$color.bars
setanalysis_defaults$show.plot.sc

#Diese Änderung ginge nicht, weil die Variable nicht existiert
#change.analysis.defaults(color.width2 = "turquoise")

}
