% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{change_settings}
\alias{change_settings}
\title{Funktion um Einstellungsvariablen anzupassen}
\usage{
change_settings(...)
}
\arguments{
\item{...}{Argumente, die die Einstellungen ändern sollen, z. B.
\code{color.bars = "red"} oder \code{show.plot.sc = FALSE}.
Die Namen der Argumente müssen mit den Namen der Einstellungsvariablen übereinstimmen.}
}
\description{
Funktion um Einstellungsvariablen anzupassen
}
\examples{
#Hier wird die Farbe der Balken auf rot geändert und eingestellt, dass keine SC-Plots gezeigt werden sollen
change_settings(color.bars = "red", show.plot.sc = FALSE)

#Eine Überprüfung zeigt, dass die Änderungen erfolgreich waren
settings$color.bars
settings$show.plot.sc

#Diese Änderung geht nicht, weil die Variable nicht existiert
change_settings(color.width2 = "turquoise")
}
