% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.label.R
\name{get.label}
\alias{get.label}
\title{Extrahiert bei MC-Fragen das Label der Antwortoption}
\usage{
get.label(x, match = ": ")
}
\arguments{
\item{x}{Eine Spalte.}

\item{match}{Die Zeichen, die direkt vor der Antwortoption stehen (üblicherweise ´: ´)}
}
\value{
Eine Zeichenkette (das Label).
}
\description{
Extrahiert bei MC-Fragen das Label der Antwortoption
}
