% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_label}
\alias{get_label}
\title{Extrahiert bei MC-Fragen das Label der Antwortoption}
\usage{
get_label(x, match = ": ")
}
\arguments{
\item{x}{Eine Spalte.}

\item{match}{Die Zeichen, die direkt vor der Antwortoption stehen (üblicherweise ´: ´)}
}
\value{
Eine Zeichenkette (das Label).
}
\description{
Extrahiert bei MC-Fragen das Label der Antwortoption
}
