% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{merge-Funktion für Schulnoten}
\usage{
grade(
  x,
  kennung,
  show.table = TRUE,
  already.aggr = FALSE,
  inkl = "nr",
  nr = ""
)
}
\arguments{
\item{x}{Daten}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{show.table}{Soll Tabelle gezeigt werden?}

\item{already.aggr}{Sind die Daten bereits aggregiert, bei TRUE wird nicht aggregiert}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}
}
\description{
merge-Funktion für Schulnoten
}
\examples{
markdown_in_viewer(grade(BspDaten$dataLVE$Note,
                                   kennung = BspDaten$dataLVE$Kennung))
}
