% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.tabelle.R
\name{input.tabelle}
\alias{input.tabelle}
\title{Aufbereiten der Berichtstabelle}
\usage{
input.tabelle(blank.path = NULL, rules.path = NULL)
}
\arguments{
\item{blank.path}{Der Dateipfad zur rohen Berichtstabelle (blank-Tabelle) als Excel-Liste. Sofern nichts angegeben wird, kann man eine Datei über ein Dialogfenster auswählen.}

\item{rules.path}{Der Dateipfad zur zugehörigen Regeltabelle (TRUE-FALSE Tabelle) als Excel-Liste. Sofern nichts angegeben wird, kann man eine Datei über ein Dialogfenster auswählen.}
}
\value{
Die aufbereitete Berichtstabelle als dataframe.
}
\description{
Diese Funktion liest die Excel-Tabellen ein, in denen einmal die zu erstellenden Berichten und die TRUE-FALSE-REgeln stehen
}
