% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv.kable.R
\name{lv.kable}
\alias{lv.kable}
\title{Funktion zur Tabellenerstellung}
\usage{
lv.kable(
  x,
  caption = NULL,
  col.width = "default",
  bold = TRUE,
  bold.col1 = TRUE,
  digits = 2,
  escape = FALSE,
  striped = TRUE
)
}
\arguments{
\item{x}{Objekt (üblicherweise Dataframe)}

\item{caption}{Titel der Tabelle}

\item{col.width}{Vektor der Spaltenbreiten, bei "default" automatische Spaltenbreiten}

\item{bold}{Sollen die Kopfzeile fettgedruckt sein?}

\item{bold.col1}{Soll die erste Zelle der Kopfzeile fettgedruckt sein?}

\item{digits}{Anzahl der Nachkommastellen}

\item{escape}{?}

\item{striped}{Soll die Tabelle Streifen (Schattierungen) erhalten}
}
\value{
Tabelle
}
\description{
Funktion zur Tabellenerstellung
}
\examples{
lv.kable(head(mtcars, 10))
}
