% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.fachsem.R
\name{merge.fachsem}
\alias{merge.fachsem}
\title{merge-Funktion für Fachsemester}
\usage{
\method{merge}{fachsem}(x, fig.height = 5, cutoff = 12, group = "a", inkl = "nr", nr = "")
}
\arguments{
\item{x}{Daten}

\item{fig.height}{Höhe des Plots im Dokument}

\item{cutoff}{cutoff-Wert, alle Werte >= cutGoff werden zusammengefasst}

\item{group}{Gruppe: "a" für alle, "b" für Bachelor und "m" für Master}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}
}
\description{
merge-Funktion für Fachsemester
}
\examples{
merge.fachsem(BspDaten$dataLVE$FachSemN) |> markdown_in_viewer()

}
