% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.mc.R
\name{merge.mc}
\alias{merge.mc}
\title{merge-Funktion für MC-Fragen}
\usage{
\method{merge}{mc}(
  x,
  head = "default",
  col1.name = "Antwortoption",
  col2.name = "n",
  show.table = TRUE,
  fig.height = "default",
  inkl = "nr",
  nr = "",
  lime = FALSE,
  filter = FALSE,
  valid.perc = TRUE,
  order.table = FALSE,
  digits = 1,
  show.plot = setanalysis_defaults$show.plot.mc
)
}
\arguments{
\item{x}{Daten (dataframe mit mehreren Spalten) -> Wichtig: Darauf achten, dass Labels enthalten sind}

\item{head}{Fragetext, bei "default" wird dieser automatisch aus den Labels gezogen}

\item{col1.name}{Name der Antwortoption in der Tabelle}

\item{col2.name}{Name der n-Spalte in der Tabelle}

\item{show.table}{Soll die Tabelle angezeigt werden?}

\item{fig.height}{Höhe der Abbildung, bei "default" ist es „Anzahl der Antwortoptionen“ * 0.75 + 1}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{lime}{Für Daten im Format nach LimeSurvey Export (nach Syntax-Skript)}

\item{filter}{FILTER-Klammer für LimeSurvey}

\item{valid.perc}{Mit gültigen Prozent?}

\item{order.table}{Soll nach Häufigkeit sortiert werden? "decreasing" für absteigendes Sortieren}

\item{digits}{Wie viele Nachkommastellen sollen angezeigt werden?}

\item{show.plot}{Soll der Plot angezeigt werden?}
}
\description{
merge-Funktion für MC-Fragen
}
\examples{

# Gewünschte Items auswählen -------------------------------------------

Abschlüsse <- dplyr::select(BspDaten$dataSHOWUP, abschluss_1:abschluss_8)

# merge.mc-Funktion ausführen und anzeigen -----------------------------

merge.mc(Abschlüsse) |> markdown_in_viewer()

}
