% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.num.R
\name{merge.num}
\alias{merge.num}
\title{Funktion für die Abbildung numerischer (quasi metrisch, wie Alter, Abschlussnote, etc.)
Kann auch für Balkendiagramme bei ordinalen Skalen genutzt werden}
\usage{
\method{merge}{num}(
  x,
  inkl = "nr",
  nr = "",
  main = "",
  xlab = "",
  cut.breaks = "",
  cut.labels = "",
  show.table = TRUE,
  fig.height = 6,
  cutoff = FALSE
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{main}{Überschrift für Plot}

\item{xlab}{Beschriftung x-Achse}

\item{cut.breaks}{Soll es cuts geben? Wo sollen die "breaks" sein, siehe cut()}

\item{cut.labels}{Soll es cuts geben? Wo sollen die "labels" sein, siehe cut()}

\item{show.table}{Soll die Tabelle gezeigt werden}

\item{fig.height}{Höhe der Abbildung}

\item{cutoff}{Soll es einen cutoff geben? Alle Werte >= cutoff werden zusammengefasst; Ist nicht mit cuts möglich!}
}
\description{
Funktion für die Abbildung numerischer (quasi metrisch, wie Alter, Abschlussnote, etc.)
Kann auch für Balkendiagramme bei ordinalen Skalen genutzt werden
}
\examples{

markdown_in_viewer(merge.num(BspDaten$dataSHOWUP$zugang_note,
                             xlab = "Durchschnittsnote für Hochschulzugangsberechtigung",
                             cut.breaks = c(0, 1.4, 1.9, 2.4, 2.9, 3.4, 2000),
                             cut.labels = c("1,0 bis 1,4","1,5 bis 1,9","2,0 bis 2,4",
                                            "2,5 bis 2,9","3,0 bis 3,4","3,5 bis 4.0")))


}
