% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.open.R
\name{merge.open}
\alias{merge.open}
\title{Funktion für offene Antworten}
\usage{
\method{merge}{open}(
  x,
  inkl = "nr",
  inkl.global = set.analysis.defaults$inkl.open,
  nr = "",
  anchor = FALSE,
  freq = "auto"
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{inkl.global}{Zweite inkl-Variable, die die globale Variable "inkl.open" abfragt. Kann auch in TRUE oder FALSE geändert werden}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{anchor}{Falls über open.answers Anker kriiert wurden hier die Nummer angeben}

\item{freq}{Sollen gleiche offene Antworten zusammengefasst werden? Dann werden auch Häufigkeiten angezeigt}
}
\description{
Funktion für offene Antworten
}
\examples{
merge.open(BspDaten$dataSHOWUP$offen, anchor = 1) |> markdown_in_viewer()

}
