% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.sc.R
\name{merge.sc}
\alias{merge.sc}
\title{merge-Funktion für single-choice Fragen}
\usage{
\method{merge}{sc}(
  x,
  inkl = "nr",
  nr = "",
  fig.height = "default",
  already.labels = FALSE,
  col2.name = "n",
  order.table = FALSE,
  show.plot = set.analysis.defaults$show.plot.sc,
  pagebreak = FALSE,
  digits = 1
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{fig.height}{Höhe der Abbildung, bei "default" ist es „Anzahl der Fragen“ * 0.75 + 1}

\item{already.labels}{Wurden die Daten bereits in Label umgewandelt?}

\item{col2.name}{Name der n-Spalte in der Tabelle}

\item{order.table}{Soll nach Häufigkeit sortiert werden? "decreasing" für absteigendes Sortieren}

\item{show.plot}{Soll der Plot angezeigt werden?}

\item{pagebreak}{Seitenumbrüche mittendrin verhindern?}

\item{digits}{Anzahl der Nachkommastellen in der Tabelle}
}
\description{
merge-Funktion für single-choice Fragen
}
\examples{
markdown_in_viewer(merge.sc(BspDaten$dataLVE$V3_D, inkl = TRUE, nr = 1))

}
