% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.wl.R
\name{merge.wl}
\alias{merge.wl}
\title{merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)}
\usage{
\method{merge}{wl}(WL, kennung, already.aggr = FALSE)
}
\arguments{
\item{WL}{WL der Daten}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{already.aggr}{Sind die Daten bereits aggregiert?}
}
\description{
merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)
}
\examples{
merge.wl(BspDaten$dataLVE$WL,
BspDaten$dataLVE$Kennung, already.aggr = FALSE) |> markdown_in_viewer()

}
