% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_open.R
\name{merge_open}
\alias{merge_open}
\title{Funktion für offene Antworten
Die Funktion war ehemals auf zwei (jetzt veraltete) Funktionen aufgeteilt:
\itemize{
\item \code{open.answers()}: Verweis auf Anhang bei Berichten mit Anhang
\item \code{merge.open()}: Eigentliche Auswertung der offenen Antworten
}}
\usage{
merge_open(
  x,
  inkl = "nr",
  inkl_global = setanalysis_defaults$inkl.open,
  nr = "",
  freq = "auto",
  appendix = setanalysis_defaults$open.appendix,
  is_appendix = FALSE,
  anchor = FALSE
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{inkl_global}{Zweite inkl-Variable, die die globale Variable "inkl.open" abfragt. Kann auch in TRUE oder FALSE geändert werden}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{freq}{Sollen gleiche offene Antworten zusammengefasst werden? Dann werden auch Häufigkeiten angezeigt.
"auto" führt zur Anzeige der Häufigkeiten, wenn Antworten mehrfach vorkommen, sonst nicht.}

\item{appendix}{Gibt es einen Extra-Anhang, in dem die offenen Antworten gesammelt werden sollen?}

\item{is_appendix}{Nur relevant, falls es einen Anhang gibt. Wenn TRUE, wird der Output für den Anhang erzeugt.}

\item{anchor}{Nur relevant, falls es einen Anhang gibt. Anker, damit auf den Output weiter oben im pdf Verlinkt werden kann.}
}
\description{
Funktion für offene Antworten
Die Funktion war ehemals auf zwei (jetzt veraltete) Funktionen aufgeteilt:
\itemize{
\item \code{open.answers()}: Verweis auf Anhang bei Berichten mit Anhang
\item \code{merge.open()}: Eigentliche Auswertung der offenen Antworten
}
}
\examples{

# Beispiel für Bericht mit Anhang – Häufigkeiten werden angezeigt
{
  merge_open(BspDaten$dataSHOWUP$offen, appendix = TRUE)
  appendix.open()
} |> markdown_in_viewer()

# Ergebnis für Bericht ohne Anhang – Ohne Häufigkeiten, weil jeder Eintrag nur einmal
merge_open(BspDaten$dataSHOWUP$offen, appendix = FALSE) |> markdown_in_viewer()

}
