% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{set.analysis.defaults}
\alias{set.analysis.defaults}
\title{Globale Umgebung für die Paketkonfiguration}
\format{
An object of class \code{environment} of length 14.
}
\usage{
set.analysis.defaults
}
\description{
Diese Umgebung wird verwendet, um Konfigurationswerte wie Farben zu speichern.
}
\keyword{datasets}
