% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{setanalysis_defaults}
\alias{setanalysis_defaults}
\title{Globale Umgebung für die Paketkonfiguration}
\usage{
setanalysis_defaults
}
\description{
Diese Umgebung wird verwendet, um Konfigurationswerte wie Farben zu speichern.
}
