% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.stat.single.R
\name{table.stat.single}
\alias{table.stat.single}
\title{Einfache Statistiktabelle für ein Item ohne Fragetext in Tabelle}
\usage{
table.stat.single(
  x,
  caption = NULL,
  md = FALSE,
  col1.name = "N_votes",
  col1.width = setanalysis_defaults$col1.width.tss,
  bold = TRUE,
  bold.col1 = TRUE
)
}
\arguments{
\item{x}{Daten}

\item{caption}{Überschrift der Tabelle}

\item{md}{Mit Median?}

\item{col1.name}{Name der ersten Zelle der Kopfzeile}

\item{col1.width}{Breite der Kopfzeile}

\item{bold}{Fettdruck der Kopfzeile}

\item{bold.col1}{Fettdruck der ersten Zelle der Kopfzeile}
}
\value{
Tabelle
}
\description{
Einfache Statistiktabelle für ein Item ohne Fragetext in Tabelle
}
