% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.alternative.R
\name{text.alternative}
\alias{text.alternative}
\title{Funktion für Ausweichoptionen}
\usage{
\method{text}{alternative}(x, show.alt = TRUE, number = 6, alt1 = FALSE, alt1.num = 0)
}
\arguments{
\item{x}{Daten}

\item{show.alt}{Zeige Ausweichoptionen, falls es sie gibt}

\item{number}{Skala (OHNE AUSWEICHOPTIONEN!)}

\item{alt1}{Text für erste Ausweichoption (standardmäßig 0 in den Daten, siehe alt1.num)}

\item{alt1.num}{0}
}
\description{
Funktion für Ausweichoptionen
}
