% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{aggr_data}
\alias{aggr_data}
\title{Testet, ob die Labels aus personalized.info so im Datensatz vorkommen}
\usage{
aggr_data(vars, kennung)
}
\arguments{
\item{vars}{Variablen (oder eine Variable), die aggregiert werden sollen}

\item{kennung}{Die Kennungen (z. B. LV-Kennungen oder Fallnummern), nach denen die Daten aggregiert werden sollen}
}
\description{
Testet, ob die Labels aus personalized.info so im Datensatz vorkommen
}
