% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendix.open.R
\name{appendix.open}
\alias{appendix.open}
\title{Funktion um alle offenen Antworten unten in den Anhang zu packen}
\usage{
appendix.open()
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{inkl.global}{Zweite inkl-Variable, die die globale Variable "inkl.open" abfragt. Kann auch in TRUE oder FALSE geändert werden}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}
}
\description{
Funktion um alle offenen Antworten unten in den Anhang zu packen
}
\examples{
# Damit diese Funktion sinnvoll funktioniert, muss vorher mindestens eine
# offene Frage aufgerufen worden
invisible(capture.output(open.answers(BspDaten$dataSHOWUP$offen)))
appendix.open() |> markdown.in.viewer()

}
