% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendix.open.R
\name{appendix.open}
\alias{appendix.open}
\title{Funktion um alle offenen Antworten unten in den Anhang zu packen}
\usage{
appendix.open()
}
\description{
Funktion um alle offenen Antworten unten in den Anhang zu packen
}
\examples{
# Damit diese Funktion sinnvoll funktioniert, muss vorher mindestens eine
# offene Frage aufgerufen worden
invisible(capture.output(open.answers(BspDaten$dataSHOWUP$offen)))
appendix.open() |> markdown.in.viewer()

}
