% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendix_open.R
\name{appendix_open}
\alias{appendix_open}
\title{Funktion um alle offenen Antworten unten in den Anhang zu packen
\code{appendix.open()} ist eine veraltete Schreibweise der gleichen Funktion}
\usage{
appendix_open(freq = "auto")
}
\arguments{
\item{freq}{Sollen die offenen Antworten nach Häufigkeit gruppiert werden?}
}
\description{
Funktion um alle offenen Antworten unten in den Anhang zu packen
\code{appendix.open()} ist eine veraltete Schreibweise der gleichen Funktion
}
\examples{
# Damit diese Funktion sinnvoll funktioniert, muss vorher mindestens eine
# offene Frage aufgerufen worden
invisible(capture.output(merge_open(BspDaten$dataSHOWUP$offen, appendix = TRUE)))
appendix_open() |> markdown_in_viewer()

}
