% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.freq.R
\name{barplot.freq}
\alias{barplot.freq}
\title{Barplot zur Abbildung von Häufigkeiten (kann auch für Balkendiagramme bei ordinalen Skalennivaus genutzt werden)}
\usage{
\method{barplot}{freq}(
  x,
  color = set.analysis.defaults$color.bars,
  xlab = "",
  main = "",
  cutoff = FALSE
)
}
\arguments{
\item{x}{Daten}

\item{color}{Farbe Balken}

\item{xlab}{Beschriftung x-Achse}

\item{main}{Titel der Abbildung}

\item{cutoff}{Cutoff,z.B. werden bei 12alle Werte über 12 als "12+" abgebildet}
}
\value{
Barplot
}
\description{
Barplot zur Abbildung von Häufigkeiten (kann auch für Balkendiagramme bei ordinalen Skalennivaus genutzt werden)
}
