% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.sc.mc.R
\name{barplot.sc.mc}
\alias{barplot.sc.mc}
\title{Barplot zur Abbildung von von SC/MC-Fragen}
\usage{
\method{barplot}{sc.mc}(x, color = set.analysis.defaults$color.bars, xlab = "")
}
\arguments{
\item{x}{Daten (data.frame mit Fragetexten, Häufigkeit und Prozent)}

\item{color}{Farbe Balken}

\item{xlab}{Beschriftung x-Achse}
}
\value{
Barplot
}
\description{
Barplot zur Abbildung von von SC/MC-Fragen
}
