% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_freq.R
\name{barplot_freq}
\alias{barplot_freq}
\title{Barplot zur Abbildung von Häufigkeiten (kann auch für Balkendiagramme bei ordinalen Skalennivaus genutzt werden)}
\usage{
barplot_freq(x, xlab = "")
}
\arguments{
\item{x}{Daten}

\item{xlab}{Achsenbeschriftung}
}
\value{
Barplot
}
\description{
Barplot zur Abbildung von Häufigkeiten (kann auch für Balkendiagramme bei ordinalen Skalennivaus genutzt werden)
}
\examples{

# Beispiel für Verwendung in merge_num ---------------------------------

barplot_freq(BspDaten$Plots$num, xlab = "Durchschnittsnote für Hochschulzugangsberechtigung")

# Beispiel für Verwendung in merge_fsem ---------------------------------

barplot_freq(BspDaten$dataLVE$FachSemN, xlab = "Fachsemester alle")

}
