% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_scmc.R
\name{barplot_scmc}
\alias{barplot_scmc}
\title{Barplot zur Abbildung von von SC/MC-Fragen}
\usage{
barplot_scmc(x, xlab = "")
}
\arguments{
\item{x}{Daten (data.frame mit Fragetexten, Häufigkeit und Prozent)}

\item{xlab}{Beschriftung x-Achse}

\item{color}{Farbe Balken}
}
\value{
Barplot
}
\description{
Barplot zur Abbildung von von SC/MC-Fragen
}
\examples{

# Beispiel für Verwendung in merge_sc ----------------------------------
barplot_scmc(BspDaten$Plots$sc, xlab = "Häufigkeit")

# Beispiel für Verwending in merge_mc ----------------------------------
barplot_scmc(BspDaten$Plots$mc, xlab = "Häufigkeit")

}
