% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_bar.R
\name{barplot_sk}
\alias{barplot_sk}
\title{Barplot-Boxplot Hypbrid für die Darstellung von ordinalskalierten Variablen
(analog zu alten EvaSys-Skalen)}
\usage{
barplot_sk(x, tmin, tmax, number = 6)
}
\arguments{
\item{x}{Daten}

\item{tmin}{Beschriftung links}

\item{tmax}{Beschriftung rechts}

\item{number}{Skala (6 für Sechserskala etc.)}
}
\value{
Barplot-Boxplot-Hybrid
}
\description{
Die optimale Chunk-Einstellung hierfür ist: fig.width = 6, fig.height = 1.4
}
\examples{
barplot_sk(BspDaten$dataSHOWUP$info_ausr_studgang,
  tmin = "stimme gar nicht zu", tmax = "stimme voll zu"
)

}
