% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.aggr.sk.R
\name{boxplot.aggr.sk}
\alias{boxplot.aggr.sk}
\title{Boxplots für Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber überarbeitet werden}
\usage{
\method{boxplot}{aggr.sk}(x, item_labels, skala)
}
\arguments{
\item{x}{Daten}

\item{item_labels}{Labels/Text/Beschriftungen der Y-Achse}

\item{skala}{Skala der x-Achse}
}
\value{
Boxplot
}
\description{
Boxplots für Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber überarbeitet werden
}
\examples{

boxplot.aggr.sk(BspDaten$Plots$aggr.data,
                BspDaten$Plots$aggr.labels,
                BspDaten$Plots$aggr.skala)

}
