% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.aggr.sk.R
\name{boxplot.aggr.sk}
\alias{boxplot.aggr.sk}
\title{Boxplots f端r Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber 端berarbeitet werden}
\usage{
\method{boxplot}{aggr.sk}(x, p, q, d, e, color = settings$color.bars)
}
\arguments{
\item{x}{daten}

\item{p}{labels/Text/Beschriftungen Y Achse}

\item{q}{Skala x Achse}

\item{d}{Anzahl von p}

\item{e}{Anzahl von q}

\item{color}{Farbe, Standard ist der Wert von color.bars}
}
\value{
Boxplot
}
\description{
Boxplots f端r Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber 端berarbeitet werden
}
