% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.workload.R
\name{boxplot.workload}
\alias{boxplot.workload}
\title{Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden}
\usage{
\method{boxplot}{workload}(x, p, q, d, e, nums)
}
\arguments{
\item{x}{Daten}

\item{p}{Text/Beschriftungen Y Achse}

\item{q}{Skala x Achse}

\item{d}{Anzahl von p}

\item{e}{Anzahl von q}

\item{nums}{?}
}
\value{
Boxplot
}
\description{
Boxplot mit Workloads der LVs: Funktioniert, sollte überarbeitet werden
}
