% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_aggr_sk.R
\name{boxplot_aggr_sk}
\alias{boxplot_aggr_sk}
\title{Boxplots für Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber überarbeitet werden}
\usage{
boxplot_aggr_sk(x, item_labels, skala)
}
\arguments{
\item{x}{Daten}

\item{item_labels}{Labels/Text/Beschriftungen der Y-Achse}

\item{skala}{Skala der x-Achse}
}
\value{
Boxplot
}
\description{
Boxplots für Skalenfragen auf aggregiertem Niveau: Funktioniert, sollte aber überarbeitet werden
}
\examples{

boxplot_aggr_sk(BspDaten$Plots$aggr.data,
                BspDaten$Plots$aggr.labels,
                BspDaten$Plots$aggr.skala)

}
