% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_rueck.R
\name{boxplot_rueck}
\alias{boxplot_rueck}
\title{Abbildung des Rücklaufs}
\usage{
boxplot_rueck(x)
}
\arguments{
\item{x}{(aggregierte) Daten für den Boxplot}
}
\value{
Boxplot der Rücklaufs
}
\description{
Abbildung des Rücklaufs
}
\examples{
boxplot_rueck(BspDaten$Plots$rueck)

}
