% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_box.R
\name{boxplot_wl}
\alias{boxplot_wl}
\title{Boxplot des Workloads}
\usage{
boxplot_wl(
  x,
  skala = c("0h", "1h", "2h", "3h", "4h", "5h", "6h", "7h", "8h", "9h", "10h", "11h",
    "12h", "mehr\\nals 12h")
)
}
\arguments{
\item{x}{Numerischer Vektor mit dem Workload je Lehrveranstaltung (Stunden
pro Woche).}

\item{skala}{Beschriftungen der x-Achse, eine pro Stufe ab 0 Stunden.}
}
\value{
Kein verwertbarer Wert (unsichtbar); die Abbildung wird gezeichnet.
}
\description{
Zeichnet einen waagerechten Boxplot des angegebenen Workloads in Stunden
pro Woche; unter der Achse steht die Anzahl der Lehrveranstaltungen. Wird
von \code{\link[=merge_wl]{merge_wl()}} verwendet.
}
\examples{
boxplot_wl(BspDaten$Plots$WL)

}
\seealso{
Weitere Grafikfunktionen:
\code{\link[=barplot_freq]{barplot_freq()}},
\code{\link[=barplot_scmc]{barplot_scmc()}},
\code{\link[=barplot_sk]{barplot_sk()}},
\code{\link[=boxplot_aggr_sk]{boxplot_aggr_sk()}},
\code{\link[=boxplot_grade]{boxplot_grade()}},
\code{\link[=boxplot_rueck]{boxplot_rueck()}}
}
\concept{grafiken}
