% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.boxplot.R
\name{bsp.boxplot}
\alias{bsp.boxplot}
\title{Beispiel-Boxplot mit Beschriftung}
\usage{
bsp.boxplot(
  x = "default",
  color = set.analysis.defaults$color.bars,
  family = set.analysis.defaults$font.family
)
}
\arguments{
\item{x}{Daten, bei "default" wird ein Beispieldatensatz genutzt}

\item{color}{Farbe des Boxplots}

\item{family}{Schriftart}
}
\value{
Beispiel-Boxplot
}
\description{
Beispiel-Boxplot mit Beschriftung
}
