% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.evasys.sk6.R
\name{bsp.evasys.sk6}
\alias{bsp.evasys.sk6}
\title{Beispiel einer Sechserskala}
\usage{
bsp.evasys.sk6(x = "default")
}
\arguments{
\item{x}{Daten, bei "default" wird ein Beispieldatensatz genutzt}
}
\value{
Grafik
}
\description{
Beispiel einer Sechserskala
}
