% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{bsp_boxplot}
\alias{bsp_boxplot}
\title{Legende: beschrifteter Beispiel-Boxplot}
\usage{
bsp_boxplot(x = "default")
}
\arguments{
\item{x}{Werte für den Boxplot. Bei \code{"default"} werden feste
Beispielwerte verwendet, zu denen die Beschriftungen passen.}
}
\value{
Nichts (unsichtbar \code{NULL}). Die Abbildung wird als Sub-Chunk in
den Bericht ausgegeben (siehe \code{\link[=subchunkify]{subchunkify()}}).
}
\description{
Gibt einen Beispiel-Boxplot aus, an dem Ausreißer, Median und Maximum
beschriftet sind. Gedacht für den Abschnitt „Erläuterung zu Grafiken“ am
Anfang eines Berichts.
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
bsp_boxplot()
\dontshow{
setwd(.old_wd)
}

}
\seealso{
Weitere Legenden:
\code{\link[=bsp_evasys_sk6]{bsp_evasys_sk6()}},
\code{\link[=bsp_table_stat]{bsp_table_stat()}}
}
\concept{legenden}
