% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{bsp_boxplot}
\alias{bsp_boxplot}
\title{Beispiel-Boxplot mit Beschriftung}
\usage{
bsp_boxplot(x = "default")
}
\arguments{
\item{x}{Daten, bei "default" wird ein Beispieldatensatz genutzt}
}
\value{
Beispiel-Boxplot
}
\description{
Beispiel-Boxplot mit Beschriftung
}
\examples{
bsp_boxplot() |> markdown_in_viewer()

}
