% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.evasys.sk6.R
\name{bsp_evasys_sk6}
\alias{bsp_evasys_sk6}
\title{Beispiel einer Sechserskala}
\usage{
bsp_evasys_sk6(x = "default")
}
\arguments{
\item{x}{Daten, bei "default" wird ein Beispieldatensatz genutzt}
}
\value{
Grafik
}
\description{
Beispiel einer Sechserskala
}
\examples{
bsp_evasys_sk6() |> markdown_in_viewer()

}
