% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BspDaten.R
\docType{data}
\name{BspDaten}
\alias{BspDaten}
\title{Beispiel-Daten}
\format{
An object of class \code{list} of length 3.
}
\usage{
BspDaten
}
\description{
Dieses Objekt enthält mehrere Beispieldaten, die in den Beispielen der
Dokumentation genutzt werden. Die Daten entstamme der LVE 2024.
}
\examples{
BspDaten$dataLVE
BspDaten$pInfo$FB.txt
}
\keyword{datasets}
