% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evasys_read_data.R
\name{evasys_read_data}
\alias{evasys_read_data}
\title{Aufbereitung von Rohdaten aus evasys}
\usage{
evasys_read_data(raw.data.path = NULL, codebook.path = NULL)
}
\arguments{
\item{raw.data.path}{Der Dateipfad zu den Rohdaten (eine csv-Datei). Sofern nichts angegeben wird, kann man eine Datei über ein Dialogfenster auswählen.}

\item{codebook.path}{Der Dateipfad zum zugehörigen Codebook aus evasys (eine csv-Datei). Sofern nichts angegeben wird, kann man eine Datei über ein Dialogfenster auswählen.}
}
\value{
Der aufbereitete Datensatz als dataframe.
}
\description{
Diese Funktion liest Rohdaten aus EvaSys ein und bereitet sie auf. Dafür wird der Rohdatensatz als csv-Datei sowie das zugehörige Codebuch als csv-Datei benötigt.
}
