% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{label_test}
\alias{label_test}
\title{Testet, ob die Labels aus personalized.info so im Datensatz vorkommen}
\usage{
label_test(col, var, exception = "alle")
}
\arguments{
\item{col}{Spalte aus der Info-Tabelle, z.B. info$Fb.text}

\item{var}{Variable aus Datensatz, die der Spalte entspricht}

\item{exception}{Ausnahmen, die nicht überprüft werden sollen}
}
\description{
Testet, ob die Labels aus personalized.info so im Datensatz vorkommen
}
\examples{
# In diesem Fall wird die Variable "FB.text" aus der Info mit der Variable
# "Teilbereich" aus dem Datensatz verglichen und alles stimmt
label_test(BspDaten$pInfo$FB.txt, BspDaten$dataLVE$Teilbereich)

# So sieht es aus, wenn die Labels nicht komplett übereinstimmen:
label_test(BspDaten$pInfo$FB.txt.falsch, BspDaten$dataLVE$Teilbereich)

}
