% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{list.open.answers}
\alias{list.open.answers}
\title{Umgebung für die offenen Antworten -------------------------------------}
\format{
An object of class \code{environment} of length 1.
}
\usage{
list.open.answers
}
\description{
Diese Umgebung wird verwendet, um die offenen Antworten für die verschiedenen Fragen zu speichern.
}
\keyword{datasets}
