% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{list_open_answers}
\alias{list_open_answers}
\title{Umgebung für die offenen Antworten -------------------------------------}
\usage{
list_open_answers
}
\description{
Diese Umgebung wird verwendet, um die offenen Antworten für die verschiedenen Fragen zu speichern.
}
