% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_table.R
\name{lv_table}
\alias{lv_table}
\title{Funktion zur Tabellenerstellung}
\usage{
lv_table(
  x,
  col.width = 1,
  bold = TRUE,
  bold.corner = TRUE,
  digits = 2,
  striped = TRUE
)
}
\arguments{
\item{x}{Objekt (üblicherweise Dataframe)}

\item{col.width}{Vektor der Spaltenbreiten, bei "default" automatische Spaltenbreiten}

\item{bold}{Sollen die Kopfzeile fettgedruckt sein?}

\item{bold.corner}{Soll der header fett sein?}

\item{digits}{Anzahl der Nachkommastellen}

\item{striped}{Soll die Tabelle Streifen (Schattierungen) erhalten}
}
\value{
Tabelle
}
\description{
Funktion zur Tabellenerstellung
}
\examples{
lv_table(head(mtcars, 10))
}
