% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_in_viewer.R
\name{markdown_in_viewer}
\alias{markdown_in_viewer}
\title{Funktion um den Markdown-Code, welcher durch eine Funktion erzeugt wurde
direkt im Viewer anzuzeigen}
\usage{
markdown_in_viewer(markdown_function)
}
\arguments{
\item{markdown_function}{Funktion, welche Markdown-Code erzeugt}
}
\value{
Vorschau im Viewer, wie das Endergebnis im Dokument aussehen würde
}
\description{
Diese Funktion ist dafür gedacht, die anderen Funktionen, die
Markdown-Code mit \code{\link[=cat]{cat()}} direkt in die Konsole drucken zu testen, sie ist
nicht für die Verwendung in einem Dokument gedacht.

\code{markdown.in.viewer()} ist eine veraltete Schreibweise der gleichen Funktion
}
\examples{
markdown_in_viewer(merge.fachsem(BspDaten$dataLVE$FachSemN))
}
