% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.evasys.sk.R
\name{merge.evasys.sk}
\alias{merge.evasys.sk}
\title{Funktion für Abbildungen analog zu alten EvaSys-Skalen}
\usage{
\method{merge}{evasys.sk}(
  x,
  inkl = "nr",
  nr = "",
  show.alt = TRUE,
  number = 6,
  alt1 = FALSE,
  alt2 = FALSE,
  alt1.num = 0,
  alt2.num = 7,
  lime = FALSE,
  lime.brackets = FALSE,
  show.plot = setanalysis_defaults$show.plot.sk
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{show.alt}{Zeige Ausweichoptionen, falls es sie gibt}

\item{number}{Anzahl Antwortoptionen des Items (OHNE AUSWEICHOPTIONEN!)}

\item{alt1}{Text für erste Ausweichoption (standardmäßig 0 in den Daten, siehe alt1.num)}

\item{alt2}{Text für zweite Ausweichoption (standardmäßig 7 in den Daten, siehe alt1.num)}

\item{alt1.num}{Welche Zahl entspricht alt1}

\item{alt2.num}{Welche Zahl entspricht alt2}

\item{lime}{Handelt es sich um exportierte LimeSurvey-Daten?}

\item{lime.brackets}{Müssen eckige Klammern um den Fragetext herum entfernt werden?}

\item{show.plot}{Zeige Plot?}
}
\description{
Funktion für Abbildungen analog zu alten EvaSys-Skalen
}
\examples{
merge.evasys.sk(BspDaten$dataSHOWUP$info_ausr_studgang) |> markdown_in_viewer()


}
