% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.multi.sk.R
\name{merge.multi.sk}
\alias{merge.multi.sk}
\title{merge-Funktion für mehrere Skalenfragen, die auf einmal dargestellt werden sollen}
\usage{
\method{merge}{multi.sk}(
  x,
  kennung,
  number = "default",
  alt1 = FALSE,
  alt2 = FALSE,
  alt1.num = 0,
  alt2.num = 7,
  nr = "",
  inkl = "nr",
  tmin = "default",
  tmid = "default",
  tmax = "default",
  show.table = TRUE,
  show.plot = TRUE,
  fig.height = "default",
  col2.name = "n",
  message = "",
  aggr = FALSE
)
}
\arguments{
\item{x}{Itemdaten (Dataframe mit mehreren Spalten)}

\item{kennung}{Objekt mit Kennungen (oder Fallnummern, nur bei Aggregierung benötigt)}

\item{number}{Anzahl Antwortoptionen der Items (OHNE AUSWEICHOPTIONEN!), wird bei "default" automatisch gezogen}

\item{alt1}{Text für erste Ausweichoption (standardmäßig 0 in den Daten, siehe alt1.num)}

\item{alt2}{Text für zweite Ausweichoption (standardmäßig 7 in den Daten, siehe alt1.num)}

\item{alt1.num}{Welche Zahl entspricht alt1}

\item{alt2.num}{Welche Zahl entspricht alt2}

\item{nr}{Nummer der ersten Frage}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{tmin}{linker Pol, bei "default" wird das Label automatisch gezogen}

\item{tmid}{mittlerer Pol (für 5er Skalen), bei "default" automatisch}

\item{tmax}{rechter Pol, "default" wie oben}

\item{show.table}{Soll die Tabelle angezeigt werden?}

\item{show.plot}{Sollen Boxplots dazu angezeigt werden?}

\item{fig.height}{Höhe der Abbildung, bei "default" ist es Anzahl der Fragen + 1}

\item{col2.name}{Titel der n-Spalte, in LVE in "N\\textsubscript{courses}" ändern}

\item{message}{Soll ein Hinweistext am Anfang erfolgen?}

\item{aggr}{Sollen Daten aggregiert werden?}
}
\description{
merge-Funktion für mehrere Skalenfragen, die auf einmal dargestellt werden sollen
}
\examples{
# Objekt erstellen, dass mehrere Fragen enthält:
KF_123 <- BspDaten$dataLVE[, c("KF_01", "KF_02", "KF_03")]

# Funktion ausführen:
markdown.in.viewer(merge.multi.sk(KF_123, number = 6, aggr = TRUE,
                   kennung = BspDaten$dataLVE$Kennung))

}
