% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.wl.R
\name{merge.wl}
\alias{merge.wl}
\title{merge-Funktion f端r den Workload (funktioniert, sollte 端berarbeitet werden)}
\usage{
\method{merge}{wl}(WL, kennung, already.aggr = FALSE)
}
\arguments{
\item{WL}{WL der Daten}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{already.aggr}{Sind die Daten bereits aggregiert?}
}
\description{
merge-Funktion f端r den Workload (funktioniert, sollte 端berarbeitet werden)
}
\examples{
merge.wl(BspDaten$dataLVE$WL,
BspDaten$dataLVE$Kennung, already.aggr = FALSE) |> markdown.in.viewer()

}
