% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.wl.R
\name{merge.wl}
\alias{merge.wl}
\title{merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)}
\usage{
\method{merge}{wl}(ECTS, WL, kennung, already.aggr = FALSE)
}
\arguments{
\item{ECTS}{ECTS der Daten}

\item{WL}{WL der Daten (ECTS und WL müssen gleiche Länge haben, d.h. entweder ist bereits beides aggregiert oder keins)}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{already.aggr}{Sind die Daten bereits aggregiert?}
}
\description{
merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)
}
