% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_auto.R
\name{merge_auto}
\alias{merge_auto}
\title{Funktion, die den automatisch Itemtyp erkennt und entsprechend auswertet}
\usage{
merge_auto(x, nr_auto = TRUE, nr = "", inkl = "nr", ...)
}
\arguments{
\item{x}{auszuwertende Daten}

\item{nr_auto}{Soll die Nummer automatisch ermittelt werden?}

\item{nr}{Manuelle Eingabemöglichkeit der Nummer}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{...}{Argumente zum „weitergeben“ in die Funktion}
}
\description{
Funktion, die den automatisch Itemtyp erkennt und entsprechend auswertet
}
