% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_grade.R
\name{merge_grade}
\alias{merge_grade}
\title{merge-Funktion für Schulnoten
\code{grade()} ist eine veraltete Schreibweise der gleichen Funktion}
\usage{
merge_grade(
  x,
  kennung,
  show.table = TRUE,
  already.aggr = FALSE,
  inkl = "nr",
  nr = ""
)
}
\arguments{
\item{x}{Daten}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{show.table}{Soll Tabelle gezeigt werden?}

\item{already.aggr}{Sind die Daten bereits aggregiert?}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}
}
\description{
merge-Funktion für Schulnoten
\code{grade()} ist eine veraltete Schreibweise der gleichen Funktion
}
\examples{
merge_grade(BspDaten$dataLVE$Note,
            kennung = BspDaten$dataLVE$Kennung) |> markdown_in_viewer()
}
