% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_many.R
\name{merge_many}
\alias{merge_many}
\title{Funktion für die automatische Auswertung mehrerer verschiedener Items unterschiedlicher Typen}
\usage{
merge_many(x, multi.sk = TRUE, nr_auto = TRUE, nr = "", inkl = "nr")
}
\arguments{
\item{x}{Ausschnitt aus dem Datensatz, der ausgewertet werden soll}

\item{multi.sk}{Sollen aufeinanderfolgende sk-Items gemeinsam ausgewertet werden?}

\item{nr_auto}{Soll die Nummer automatisch ermittelt werden?}

\item{nr}{Manuelle Eingabemöglichkeit der Nummer}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}
}
\description{
Funktion für die automatische Auswertung mehrerer verschiedener Items unterschiedlicher Typen
}
