% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rueck.R
\name{merge_rueck}
\alias{merge_rueck}
\title{Boxplot für Rücklaufsabbildung (LVE)}
\usage{
merge_rueck(x, kennung)
}
\arguments{
\item{x}{Objekt mit Teilnehmendenzahlen}

\item{kennung}{Kennungen}
}
\value{
Boxplot der Rückläufe
}
\description{
Boxplot für Rücklaufsabbildung (LVE)
}
\examples{
merge_rueck(BspDaten$dataLVE$Teilnehmer,
            BspDaten$dataLVE$Kennung) |> markdown_in_viewer()

}
