% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_subj.R
\name{merge_subj}
\alias{merge_subj}
\title{merge-Funktion zum Zusammenfügen der single-choice Fragen nach dem 1. und 2. Fach}
\usage{
merge_subj(x1, x2, inkl1 = "nr1", inkl2 = "nr2", nr1 = "", nr2 = "")
}
\arguments{
\item{x1}{Daten von Fach 1}

\item{x2}{Daten von Fach 2}

\item{inkl1}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr1" zieht sich automatisch die entsprechende inkl. Variable}

\item{inkl2}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr2" zieht sich automatisch die entsprechende inkl. Variable}

\item{nr1}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{nr2}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}
}
\description{
merge-Funktion zum Zusammenfügen der single-choice Fragen nach dem 1. und 2. Fach
}
\examples{

merge_subj(BspDaten$dataSHOWUP$fach1_2FB,
           BspDaten$dataSHOWUP$fach2_2FB) |> markdown_in_viewer()

}
