% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_wl.R
\name{merge_wl}
\alias{merge_wl}
\title{merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)
\code{merge.wl()} ist eine veraltete Schreibweise der gleichen Funktion}
\usage{
merge_wl(WL, kennung, already.aggr = FALSE)
}
\arguments{
\item{WL}{WL der Daten}

\item{kennung}{Kennung/Fallnummer zum Aggregieren}

\item{already.aggr}{Sind die Daten bereits aggregiert?}
}
\description{
merge-Funktion für den Workload (funktioniert, sollte überarbeitet werden)
\code{merge.wl()} ist eine veraltete Schreibweise der gleichen Funktion
}
\examples{
merge_wl(BspDaten$dataLVE$WL, already.aggr = FALSE
         BspDaten$dataLVE$Kennung) |> markdown_in_viewer()

}
