% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.answers.R
\name{open.answers}
\alias{open.answers}
\title{Funktion für Fragen mit offenem Antwortformat (benötigt u.a. "list.open", wird am Anfang des Skripts erstellt)}
\usage{
\method{open}{answers}(
  x,
  inkl = "nr",
  inkl.global = inkl.open,
  nr = "",
  freq = FALSE,
  no.pagebreak = TRUE
)
}
\arguments{
\item{x}{Daten}

\item{inkl}{TRUE oder FALSE, ob die Funktion ausgeführt wird; "nr" zieht sich automatisch die entsprechende inkl. Variable}

\item{inkl.global}{Zweite inkl-Variable, die die globale Variable "inkl.open" abfragt. Kann auch in TRUE oder FALSE geändert werden}

\item{nr}{Nummer, die Grundlage für entsprechende inkl. Variable ist und vorne an den Fragetext gestellt wird}

\item{freq}{Sollen gleiche offene Antworten zusammengefasst werden? Dann werden auch Häufigkeiten angezeigt}

\item{no.pagebreak}{Seitenumbrüche mittendrin verhindern?}
}
\description{
Funktion für Fragen mit offenem Antwortformat (benötigt u.a. "list.open", wird am Anfang des Skripts erstellt)
}
