% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veraltete_funktionen.R
\name{setanalysis-deprecated}
\alias{setanalysis-deprecated}
\alias{appendix.open}
\alias{grade}
\alias{markdown.in.viewer}
\alias{open.answers}
\alias{boxplot.ruecklauf}
\alias{merge.evasys.sk}
\alias{merge.fachsem}
\alias{merge.mc}
\alias{merge.multi.sk}
\alias{merge.num}
\alias{merge.open}
\alias{merge.sc}
\alias{merge.subj}
\alias{merge.wl}
\title{Veraltete Funktionsnamen}
\value{
Siehe die jeweils aktuelle Funktion.
}
\description{
Diese Funktionen sind ältere Schreibweisen aktueller Funktionen. Sie
verhalten sich genau wie die jeweils aktuelle Funktion und werden nur noch
bereitgestellt, damit bestehende Berichtsvorlagen weiter funktionieren. In
neuem Code sollten die aktuellen Namen verwendet werden.\tabular{ll}{
   Veralteter Name \tab Aktuelle Funktion \cr
   \code{appendix.open()} \tab \code{\link[=appendix_open]{appendix_open()}} \cr
   \code{boxplot.ruecklauf()} \tab \code{\link[=merge_rueck]{merge_rueck()}} \cr
   \code{grade()} \tab \code{\link[=merge_grade]{merge_grade()}} \cr
   \code{markdown.in.viewer()} \tab \code{\link[=markdown_in_viewer]{markdown_in_viewer()}} \cr
   \code{merge.evasys.sk()} \tab \code{\link[=merge_sk]{merge_sk()}} \cr
   \code{merge.fachsem()} \tab \code{\link[=merge_fachsem]{merge_fachsem()}} \cr
   \code{merge.mc()} \tab \code{\link[=merge_mc]{merge_mc()}} \cr
   \code{merge.multi.sk()} \tab \code{\link[=merge_aggr_sk]{merge_aggr_sk()}} \cr
   \code{merge.num()} \tab \code{\link[=merge_num]{merge_num()}} \cr
   \code{merge.open()} \tab \code{\link[=merge_open]{merge_open()}} \cr
   \code{merge.sc()} \tab \code{\link[=merge_sc]{merge_sc()}} \cr
   \code{merge.subj()} \tab \code{\link[=merge_subj]{merge_subj()}} \cr
   \code{merge.wl()} \tab \code{\link[=merge_wl]{merge_wl()}} \cr
   \code{open.answers()} \tab \code{\link[=merge_open]{merge_open()}} mit \code{appendix = TRUE} \cr
}
}
\details{
Namen wie \code{merge.sc()} oder \code{boxplot.ruecklauf()} sehen für R wie
S3-Methoden der generischen Funktionen \code{\link[=merge]{merge()}} bzw. \code{\link[=boxplot]{boxplot()}} aus. Damit
sie nicht als solche behandelt werden, haben sie die Argumente der
generischen Funktion und leiten den Aufruf unverändert an die aktuelle
Funktion weiter. Argumente werden dabei genau so zugeordnet wie bei einem
direkten Aufruf der aktuellen Funktion.
}
\keyword{internal}
