% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veraltete_funktionen.R, R/zzz.R
\name{setanalysis-deprecated}
\alias{setanalysis-deprecated}
\alias{appendix.open}
\alias{bsp.boxplot}
\alias{bsp.evasys.sk6}
\alias{bsp.table.stat}
\alias{change.analysis.defaults}
\alias{evasys.read.data}
\alias{grade}
\alias{input.tabelle}
\alias{label.test}
\alias{markdown.in.viewer}
\alias{open.answers}
\alias{table.freq}
\alias{table.stat.multi}
\alias{table.stat.single}
\alias{boxplot.ruecklauf}
\alias{merge.evasys.sk}
\alias{merge.fachsem}
\alias{merge.mc}
\alias{merge.multi.sk}
\alias{merge.num}
\alias{merge.open}
\alias{merge.sc}
\alias{merge.subj}
\alias{merge.wl}
\alias{list.open.answers}
\title{Veraltete Funktionsnamen}
\value{
Siehe die jeweils aktuelle Funktion.
}
\description{
Diese Funktionen sind ältere Schreibweisen aktueller Funktionen. Sie
verhalten sich genau wie die jeweils aktuelle Funktion und werden nur noch
bereitgestellt, damit bestehende Berichtsvorlagen weiter funktionieren. In
neuem Code sollten die aktuellen Namen verwendet werden.\tabular{ll}{
   Veralteter Name \tab Aktueller Name \cr
   \code{appendix.open()} \tab \code{\link[=appendix_open]{appendix_open()}} \cr
   \code{boxplot.ruecklauf()} \tab \code{\link[=merge_rueck]{merge_rueck()}} \cr
   \code{bsp.boxplot()} \tab \code{\link[=bsp_boxplot]{bsp_boxplot()}} \cr
   \code{bsp.evasys.sk6()} \tab \code{\link[=bsp_evasys_sk6]{bsp_evasys_sk6()}} \cr
   \code{bsp.table.stat()} \tab \code{\link[=bsp_table_stat]{bsp_table_stat()}} \cr
   \code{change.analysis.defaults()} \tab \code{\link[=change_analysis_defaults]{change_analysis_defaults()}} \cr
   \code{evasys.read.data()} \tab \code{\link[=evasys_read_data]{evasys_read_data()}} \cr
   \code{grade()} \tab \code{\link[=merge_grade]{merge_grade()}} \cr
   \code{input.tabelle()} \tab \code{\link[=input_tabelle]{input_tabelle()}} \cr
   \code{label.test()} \tab \code{\link[=label_test]{label_test()}} \cr
   \code{list.open.answers} \tab \link{list_open_answers} \cr
   \code{markdown.in.viewer()} \tab \code{\link[=markdown_in_viewer]{markdown_in_viewer()}} \cr
   \code{merge.evasys.sk()} \tab \code{\link[=merge_sk]{merge_sk()}} \cr
   \code{merge.fachsem()} \tab \code{\link[=merge_fachsem]{merge_fachsem()}} \cr
   \code{merge.mc()} \tab \code{\link[=merge_mc]{merge_mc()}} \cr
   \code{merge.multi.sk()} \tab \code{\link[=merge_aggr_sk]{merge_aggr_sk()}} \cr
   \code{merge.num()} \tab \code{\link[=merge_num]{merge_num()}} \cr
   \code{merge.open()} \tab \code{\link[=merge_open]{merge_open()}} \cr
   \code{merge.sc()} \tab \code{\link[=merge_sc]{merge_sc()}} \cr
   \code{merge.subj()} \tab \code{\link[=merge_subj]{merge_subj()}} \cr
   \code{merge.wl()} \tab \code{\link[=merge_wl]{merge_wl()}} \cr
   \code{open.answers()} \tab \code{\link[=merge_open]{merge_open()}} mit \code{appendix = TRUE} \cr
   \code{table.freq()} \tab \code{\link[=table_freq]{table_freq()}} \cr
   \code{table.stat.multi()} \tab \code{\link[=table_stat_multi]{table_stat_multi()}} \cr
   \code{table.stat.single()} \tab \code{\link[=table_stat_single]{table_stat_single()}} \cr
}
}
\details{
Die veralteten Namen leiten jeden Aufruf unverändert an die aktuelle
Funktion weiter. Argumente werden dabei genau so zugeordnet wie bei einem
direkten Aufruf der aktuellen Funktion.

Namen wie \code{merge.sc()} oder \code{boxplot.ruecklauf()} sehen für R wie
S3-Methoden der generischen Funktionen \code{\link[=merge]{merge()}} bzw. \code{\link[=boxplot]{boxplot()}} aus. Sie
haben deshalb die Argumente der generischen Funktion, damit sie nicht als
fehlerhafte Methoden gelten.

\code{list.open.answers} ist ein zweiter Name für dieselbe Umgebung wie
\link{list_open_answers}.
}
\keyword{internal}
