% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{subchunkify}
\alias{subchunkify}
\title{Erzeugt Subchunks für die Berichte}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5, hide = FALSE)
}
\arguments{
\item{g}{Code (kann auch mit Aufzählung ("c(...)") benutzt werden)}

\item{fig_height}{Höhe des Sub-Chunks}

\item{fig_width}{Breite des Sub-Chunks}

\item{hide}{Soll für die results-Option des Chunks "asis" verwendet werden?}
}
\value{
Subchunk
}
\description{
Erzeugt Subchunks für die Berichte
}
