% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.freq.R
\name{table.freq}
\alias{table.freq}
\title{Einfache Häufigkeitstabelle}
\usage{
table.freq(
  x,
  caption = NULL,
  cutoff = FALSE,
  show.all = TRUE,
  col1.name = "",
  col2.name = "N",
  col.width = "default",
  order.table = FALSE,
  bold = TRUE,
  bold.col1 = TRUE
)
}
\arguments{
\item{x}{Daten}

\item{caption}{Überschrift der Tabelle (siehe lv.kable)}

\item{cutoff}{Soll es einen "cutoff" geben? z.B. werden bei 12 alle Werte >= 12 in "12 oder höher" dargestellt}

\item{show.all}{Sollen auch nicht gewählte Antwortoptionen angezeigt werden?}

\item{col1.name}{Name der ersten Zelle der Kopfzeile}

\item{col2.name}{Name der zweiten Zelle der Kopfzeile (standardmäßig N)}

\item{col.width}{Spaltenbreite (siehe lv.kable)}

\item{order.table}{Soll nach Häufigkeit sortiert werden? "decreasing" für absteigendes Sortieren}

\item{bold}{Soll die Kopfzeile fett sein? (siehe lv.kable)}

\item{bold.col1}{Soll die erste Zelle der Kopfzeile fett sein? (siehe lv.kable)}
}
\value{
Tabelle
}
\description{
Einfache Häufigkeitstabelle
}
