% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.stat.multi.R
\name{table.stat.multi}
\alias{table.stat.multi}
\title{Einfache Statistiktabelle für mehrere Items mit Fragetexten}
\usage{
table.stat.multi(
  x,
  col1.name = "Item",
  col2.name = "N_votes",
  alt1 = FALSE,
  alt2 = FALSE,
  alt1.list = NULL,
  alt2.list = NULL,
  digits = 2,
  bold = TRUE,
  bold.corner = TRUE,
  labels = "labels"
)
}
\arguments{
\item{x}{Daten}

\item{col1.name}{Name der ersten Zelle der Kopfzeile}

\item{col2.name}{Name der zweiten Zelle der Kopfzeile}

\item{alt1}{Text für erste Ausweichoption}

\item{alt2}{Text für zweite Ausweichoption}

\item{alt1.list}{Antworthäufigkeiten erste Ausweichoption}

\item{alt2.list}{Antworthäufigkeiten zweite Ausweichoption}

\item{digits}{Anzahl der Nachkommastellend}

\item{bold}{Sollen die Kopfzeile fettgedruckt sein? (siehe lv_table)}

\item{bold.corner}{Soll die Zelle ganz links in der Kopfzeile fettgedruckt sein?}

\item{labels}{Fragetexte, bei "labels" werden die Labels der Variablen genommen}
}
\value{
Tabelle
}
\description{
Einfache Statistiktabelle für mehrere Items mit Fragetexten
}
\examples{

table.stat.multi(BspDaten$Tabellen$multi)

}
