% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.stat.single.R
\name{table.stat.single}
\alias{table.stat.single}
\title{Einfache Statistiktabelle für ein Item ohne Fragetext in Tabelle}
\usage{
table.stat.single(
  x,
  md = FALSE,
  col1.name = "N_votes",
  bold = TRUE,
  digits = 2
)
}
\arguments{
\item{x}{Daten}

\item{md}{Mit Median?}

\item{col1.name}{Name der ersten Zelle der Kopfzeile}

\item{bold}{Fettdruck der Kopfzeile}

\item{digits}{Anzahl der Nachkommastellen in der Tabelle}
}
\value{
Tabelle
}
\description{
Einfache Statistiktabelle für ein Item ohne Fragetext in Tabelle
}
