% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_freq}
\alias{table_freq}
\title{Einfache Häufigkeitstabelle}
\usage{
table_freq(
  x,
  cutoff = FALSE,
  show.all = TRUE,
  col1.name = "",
  col2.name = "n",
  col.width = "default",
  order.table = FALSE,
  bold = TRUE,
  digits = 1
)
}
\arguments{
\item{x}{Daten}

\item{cutoff}{Soll es einen "cutoff" geben? z.B. werden bei 12 alle Werte >= 12 in "12 oder höher" dargestellt}

\item{show.all}{Sollen auch nicht gewählte Antwortoptionen angezeigt werden?}

\item{col1.name}{Name der ersten Zelle der Kopfzeile}

\item{col2.name}{Name der zweiten Zelle der Kopfzeile (standardmäßig N)}

\item{col.width}{Spaltenbreiten (siehe lv_table)}

\item{order.table}{Soll nach Häufigkeit sortiert werden? "decreasing" für absteigendes Sortieren}

\item{bold}{Soll die Kopfzeile fett sein? (siehe lv_table)}

\item{digits}{Nachkommastellen}
}
\value{
Tabelle
}
\description{
Einfache Häufigkeitstabelle
}
\examples{
table_freq(BspDaten$Tabellen$freq)

}
